waiting for trip to Finish my last Two Methods

+++ src/main/java/com/example/mini2/controllers/PaymentController.java
@@ -3,6 +3,7 @@ package com.example.mini2.controllers;
 import com.example.mini2.models.Payment;
 import com.example.mini2.services.PaymentService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.rest.webmvc.ResourceNotFoundException;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -18,44 +19,47 @@ public class PaymentController {
         this.paymentService = paymentService;
     }
 
-    // 9.4.2.1 Add Payment
     @PostMapping("/addPayment")
     public Payment addPayment(@RequestBody Payment payment) {
         return paymentService.addPayment(payment);
     }
 
-    // 9.4.2.2 Get All Payments
     @GetMapping("/allPayments")
     public List<Payment> getAllPayments() {
         return paymentService.getAllPayments();
     }
 
-    // 9.4.2.3 Get Specific Payment
     @GetMapping("/{id}")
     public Payment getPaymentById(@PathVariable Long id) {
-        return paymentService.getPaymentById(id);
+        Payment payment = paymentService.getPaymentById(id);
+        if (payment == null) {
+            throw new ResourceNotFoundException("Payment with ID " + id + " not found.");
+        }
+        return payment;
     }
-
-    // 9.4.2.4 Update Payment
     @PutMapping("/update/{id}")
     public Payment updatePayment(@PathVariable Long id, @RequestBody Payment payment) {
+        Payment existingPayment = paymentService.getPaymentById(id);
+        if (existingPayment == null) {
+            throw new ResourceNotFoundException(" payment not found");
+        }
         return paymentService.updatePayment(id, payment);
-    }
-
-    // 9.4.2.5 Delete Payment
+    } // sebk mnha b3den
     @DeleteMapping("/delete/{id}")
     public String deletePayment(@PathVariable Long id) {
+        if (paymentService.getPaymentById(id) == null) {
+            return "Payment with ID " + id + " not found.";
+        }
         paymentService.deletePayment(id);
         return "Payment with ID " + id + " has been deleted.";
     }
 
     // 9.4.2.6 Find Payments By Trip ID
-    @GetMapping("/findByTripId")
-    public List<Payment> findPaymentsByTripId(@RequestParam Long tripId) {
-        return paymentService.findPaymentsByTripId(tripId);
-    }
+    //@GetMapping("/findByTripId")
+   // public List<Payment> findPaymentsByTripId(@RequestParam Long tripId) {
+ //       return paymentService.findPaymentsByTripId(tripId);
+   // }
 
-    // 9.4.2.7 Find Payments With an Amount Greater Than a Threshold
     @GetMapping("/findByAmountThreshold")
     public List<Payment> findPaymentsWithAmountGreaterThan(@RequestParam Double threshold) {
         return paymentService.findByAmountThreshold(threshold);
